fe2o3_social: - Downgraded some types from usize to u32 and f64 to f32. fe2o3_jdat: - Created more flexible enum_getter_numeric! macro for automatic coercion of numeric values via FromDatMap.

+++ fe2o3_core/src/mem.rs
@@ -14,14 +14,14 @@ impl<T: Default> Extract for T {
 /// 
 /// Returns the current process memory usage (VmRSS) in megabytes.
 /// On non-Linux systems or if unable to read memory stats, returns 0.0.
-pub fn get_memory_usage_mb() -> f64 {
+pub fn get_memory_usage_mb() -> f32 {
     // Simple approach: read from /proc/self/status on Linux.
     if let Ok(status) = std::fs::read_to_string("/proc/self/status") {
         for line in status.lines() {
             if line.starts_with("VmRSS:") {
                 if let Some(kb_str) = line.split_whitespace().nth(1) {
                     if let Ok(kb) = kb_str.parse::<u64>() {
-                        return kb as f64 / 1024.0; // Convert KB to MB.
+                        return kb as f32 / 1024.0; // Convert KB to MB.
                     }
                 }
             }
